% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_dplyr.R
\name{browse_dplyr}
\alias{browse_dplyr}
\title{Browse the dplyr documentation}
\usage{
browse_dplyr(x = NULL)
}
\arguments{
\item{x}{Character(1). Documentation file name.}
}
\value{
Open browser.
}
\description{
Browse the dplyr documentation
}
