% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_dplyr.R
\name{browse_magrittr}
\alias{browse_magrittr}
\title{Browse the magrittr documentation}
\usage{
browse_magrittr(x = NULL)
}
\arguments{
\item{x}{Character(1). Documentation file name.}
}
\value{
Open browser.
}
\description{
Browse the magrittr documentation
}
