% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_dplyr.R
\name{browse_tidyr}
\alias{browse_tidyr}
\title{Browse the tidyr documentation}
\usage{
browse_tidyr(x = NULL)
}
\arguments{
\item{x}{Character(1). Documentation file name.}
}
\value{
Open browser.
}
\description{
Browse the tidyr documentation
}
